requirements.txt

email-validator>=2.2.0
flask-socketio>=5.5.1
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
openai
openai
openai
eventlet
python-dotenv
python-dotenv
